Adjusting The getTypesDex
Needed to adjust this file in order to add our custom type Cosmic into the calc. All I did was commented out the reference to the Type class from @smogon/calc, and rebuilded that reference class in the file and modified it. I can't test this on my local machine because I have a higher node version, but my vps can run it.